requirements.txt
asgiref==3.6.0
Django==3.2.4
django-cors-headers==3.7.0
djangorestframework==3.12.4
pytz==2022.7.1
sqlparse==0.4.3
tzdata==2022.7
